Refactor compute_output_transform; refactor wcs_from_footprints to use gwcs obj
